refactor: Extract FileUploadFactory following SRP and reduce complexity
Major refactoring to address PHPMD complexity issues and improve design:

**FileUploadFactory (new)**
- Extract all file upload logic from InputQuery into dedicated factory
- Follow Single Responsibility Principle for file upload creation
- Support service locator pattern for testing (pre-created FileUpload objects)
- Independently usable in BEAR.Resource and other contexts
- Handle mixed type parameters with InputFile attribute
- Comprehensive documentation of service locator pattern usage

**InputQuery (simplified)**
- Reduce complexity by delegating file upload logic to FileUploadFactory
- Remove 200+ lines of file upload methods
- Maintain same public API for backward compatibility
- Use FileUploadFactory for all file upload operations

**Type System Enhancement**
- Introduce Query domain type in InputQueryInterface
- Import Query type across components for consistency
- Replace generic array types with semantic Query type
- Improve type safety and code documentation

**Benefits**
- PHPMD complexity now below threshold (was 108, now <100)
- Better separation of concerns (file vs general parameter handling)
- Enhanced testability through service locator pattern
- Reusable FileUploadFactory for standalone usage
- Clearer type definitions with Query domain type
