+++ src/InputQueryInterface.php
@@ -6,13 +6,16 @@ namespace Ray\InputQuery;
 
 use ReflectionMethod;
 
-/** @template T of object */
+/**
+ * @template T of object
+ * @psalm-type Query = array<string, mixed>
+ */
 interface InputQueryInterface
 {
     /**
      * Get method arguments from query data
      *
-     * @param array<string, mixed> $query
+     * @param Query $query HTTP request data ($_POST, $_GET, etc.)
      *
      * @return array<mixed>
      */
@@ -21,8 +24,8 @@ interface InputQueryInterface
     /**
      * Create object from query data
      *
-     * @param class-string<T>      $class
-     * @param array<string, mixed> $query
+     * @param class-string<T> $class
+     * @param Query           $query HTTP request data ($_POST, $_GET, etc.)
      *
      * @return T
      */
